orca_whirlpools_docs

Module orca_whirlpools_client

Orca Whirlpools Client SDK

Overview

This package provides developers with low-level functionalities for interacting with the Whirlpool Program on Solana. It serves as a foundational tool that allows developers to manage and integrate detailed operations into their Rust projects, particularly those related to Orca’s Whirlpool Program. This package offers granular control for advanced use cases.

Compatibiltiy

Compatible with anchor versions ^0.26 but <0.30. If you enable the anchor feature of orca_whirlpools_client in cargo.toml while using a version of anchor that’s ^0.30 in your project, you may need to apply a lockfile patch to switch to a lower version:

cargo update anchor:<current-version> --precise 0.29

Compatible with solana-program versions ^1.18.0 but <3.0.0. By default, Cargo will install the latest version of Solana SDK ^v2. This can cause dependcy issues when using older versions. To solve this you can apply a lockfile patch with the following command:

cargo update solana-program:<current-version> --precise <required-version>

NOTE: if you are dealing with compatibility issues for both anchor and solana-program, the order of the patches matters. First patch anchor, then patch solana-program.

Key Features

Codama Client: The package includes a set of generated client code based on the Whirlpool Program IDL. This ensures all the necessary program information is easily accessible in a structured format. It handles all decoding and encoding of instructions and account data, making it much easier to interact with the program.

PDA (Program Derived Addresses) Utilities: This feature contains utility functions that help derive Program Derived Addresses (PDAs) for accounts within the Whirlpool Program, simplifying address derivation for developers.

Installation

cargo add orca_whirlpools_client

Usage

Here are some basic examples of how to use the package:

Deriving a PDA

To derive a PDA for a Whirlpool account, you can use the get_whirlpool_address PDA utility.

use orca_whirlpools_client::get_whirlpool_address;
use solana_sdk::pubkey::Pubkey;
use std::str::FromStr;

fn main() {
    let whirlpool_config_address = Pubkey::from_str("FcrweFY1G9HJAHG5inkGB6pKg1HZ6x9UC2WioAfWrGkR").unwrap();
    let token_mint_a = Pubkey::from_str("So11111111111111111111111111111111111111112").unwrap(); // wSOL
    let token_mint_b = Pubkey::from_str("BRjpCHtyQLNCo8gqRUr8jtdAj5AjPYQaoqbvcZiHok1k").unwrap(); // DevUSDC
    let tick_spacing = 64;

    let (whirlpool_pda, _bump) = get_whirlpool_address(&whirlpool_config_address, &token_mint_a, &token_mint_b, tick_spacing).unwrap();
    println!("{:?}", whirlpool_pda);
}

Example: Initialize Pool Instruction

The following example demonstrates how to create an InitializePool instruction:

use orca_whirlpools_client::{
    instructions::InitializePoolV2Builder,
    get_fee_tier_address,
    get_token_badge_address,
    get_whirlpool_address,
};
use solana_sdk::{
    pubkey::Pubkey,
    signer::{keypair::Keypair, Signer},
};
use std::str::FromStr;

fn main() {
    let whirlpool_config_address = Pubkey::from_str("FcrweFY1G9HJAHG5inkGB6pKg1HZ6x9UC2WioAfWrGkR").unwrap();
    let token_mint_a = Pubkey::from_str("So11111111111111111111111111111111111111112").unwrap(); // wSOL
    let token_mint_b = Pubkey::from_str("BRjpCHtyQLNCo8gqRUr8jtdAj5AjPYQaoqbvcZiHok1k").unwrap(); // DevUSDC
    let (token_badge_a, _bump) = get_token_badge_address(&whirlpool_config_address, &token_mint_a).unwrap();
    let (token_badge_b, _bump) = get_token_badge_address(&whirlpool_config_address, &token_mint_b).unwrap();
    let wallet = Keypair::new(); // CAUTION: this wallet is not persistent
    let tick_spacing = 8;
    let (whirlpool_pda, _bump) = get_whirlpool_address(&whirlpool_config_address, &token_mint_a, &token_mint_b, tick_spacing).unwrap();
    let token_vault_a = Keypair::new();
    let token_vault_b = Keypair::new();
    let (fee_tier, _bump) = get_fee_tier_address(&whirlpool_config_address, tick_spacing).unwrap();
    let token_program_a = Pubkey::from_str("TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA").unwrap();
    let token_program_b = Pubkey::from_str("TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA").unwrap();
    let initial_sqrt_price = 7459106261056563200u128;    

    let initialize_pool_v2_instruction = InitializePoolV2Builder::new()
        .whirlpools_config(whirlpool_config_address)
        .token_mint_a(token_mint_a)
        .token_mint_b(token_mint_b)
        .token_badge_a(token_badge_a)
        .token_badge_b(token_badge_b)
        .funder(wallet.pubkey())
        .whirlpool(whirlpool_pda)
        .token_vault_a(token_vault_a.pubkey())
        .token_vault_b(token_vault_b.pubkey())
        .fee_tier(fee_tier)
        .token_program_a(token_program_a)
        .token_program_b(token_program_b)
        .tick_spacing(tick_spacing)
        .initial_sqrt_price(initial_sqrt_price)
        .instruction();

    println!("{:?}", initialize_pool_v2_instruction);
}

Structs

AdaptiveFeeConstants

AdaptiveFeeTier

AdaptiveFeeVariables

CloseBundledPosition

Accounts.

CloseBundledPositionBuilder

Instruction builder for CloseBundledPosition.

CloseBundledPositionCpi

close_bundled_position CPI instruction.

CloseBundledPositionCpiAccounts

close_bundled_position CPI accounts.

CloseBundledPositionCpiBuilder

Instruction builder for CloseBundledPosition via CPI.

CloseBundledPositionInstructionArgs

CloseBundledPositionInstructionData

ClosePosition

Accounts.

ClosePositionBuilder

Instruction builder for ClosePosition.

ClosePositionCpi

close_position CPI instruction.

ClosePositionCpiAccounts

close_position CPI accounts.

ClosePositionCpiBuilder

Instruction builder for ClosePosition via CPI.

ClosePositionInstructionData

ClosePositionWithTokenExtensions

Accounts.

ClosePositionWithTokenExtensionsBuilder

Instruction builder for ClosePositionWithTokenExtensions.

ClosePositionWithTokenExtensionsCpi

close_position_with_token_extensions CPI instruction.

ClosePositionWithTokenExtensionsCpiAccounts

close_position_with_token_extensions CPI accounts.

ClosePositionWithTokenExtensionsCpiBuilder

Instruction builder for ClosePositionWithTokenExtensions via CPI.

ClosePositionWithTokenExtensionsInstructionData

CollectFees

Accounts.

CollectFeesBuilder

Instruction builder for CollectFees.

CollectFeesCpi

collect_fees CPI instruction.

CollectFeesCpiAccounts

collect_fees CPI accounts.

CollectFeesCpiBuilder

Instruction builder for CollectFees via CPI.

CollectFeesInstructionData

CollectFeesV2

Accounts.

CollectFeesV2Builder

Instruction builder for CollectFeesV2.

CollectFeesV2Cpi

collect_fees_v2 CPI instruction.

CollectFeesV2CpiAccounts

collect_fees_v2 CPI accounts.

CollectFeesV2CpiBuilder

Instruction builder for CollectFeesV2 via CPI.

CollectFeesV2InstructionArgs

CollectFeesV2InstructionData

CollectProtocolFees

Accounts.

CollectProtocolFeesBuilder

Instruction builder for CollectProtocolFees.

CollectProtocolFeesCpi

collect_protocol_fees CPI instruction.

CollectProtocolFeesCpiAccounts

collect_protocol_fees CPI accounts.

CollectProtocolFeesCpiBuilder

Instruction builder for CollectProtocolFees via CPI.

CollectProtocolFeesInstructionData

CollectProtocolFeesV2

Accounts.

CollectProtocolFeesV2Builder

Instruction builder for CollectProtocolFeesV2.

CollectProtocolFeesV2Cpi

collect_protocol_fees_v2 CPI instruction.

CollectProtocolFeesV2CpiAccounts

collect_protocol_fees_v2 CPI accounts.

CollectProtocolFeesV2CpiBuilder

Instruction builder for CollectProtocolFeesV2 via CPI.

CollectProtocolFeesV2InstructionArgs

CollectProtocolFeesV2InstructionData

CollectReward

Accounts.

CollectRewardBuilder

Instruction builder for CollectReward.

CollectRewardCpi

collect_reward CPI instruction.

CollectRewardCpiAccounts

collect_reward CPI accounts.

CollectRewardCpiBuilder

Instruction builder for CollectReward via CPI.

CollectRewardInstructionArgs

CollectRewardInstructionData

CollectRewardV2

Accounts.

CollectRewardV2Builder

Instruction builder for CollectRewardV2.

CollectRewardV2Cpi

collect_reward_v2 CPI instruction.

CollectRewardV2CpiAccounts

collect_reward_v2 CPI accounts.

CollectRewardV2CpiBuilder

Instruction builder for CollectRewardV2 via CPI.

CollectRewardV2InstructionArgs

CollectRewardV2InstructionData

DecodedAccount

DecreaseLiquidity

Accounts.

DecreaseLiquidityBuilder

Instruction builder for DecreaseLiquidity.

DecreaseLiquidityCpi

decrease_liquidity CPI instruction.

DecreaseLiquidityCpiAccounts

decrease_liquidity CPI accounts.

DecreaseLiquidityCpiBuilder

Instruction builder for DecreaseLiquidity via CPI.

DecreaseLiquidityInstructionArgs

DecreaseLiquidityInstructionData

DecreaseLiquidityV2

Accounts.

DecreaseLiquidityV2Builder

Instruction builder for DecreaseLiquidityV2.

DecreaseLiquidityV2Cpi

decrease_liquidity_v2 CPI instruction.

DecreaseLiquidityV2CpiAccounts

decrease_liquidity_v2 CPI accounts.

DecreaseLiquidityV2CpiBuilder

Instruction builder for DecreaseLiquidityV2 via CPI.

DecreaseLiquidityV2InstructionArgs

DecreaseLiquidityV2InstructionData

DeletePositionBundle

Accounts.

DeletePositionBundleBuilder

Instruction builder for DeletePositionBundle.

DeletePositionBundleCpi

delete_position_bundle CPI instruction.

DeletePositionBundleCpiAccounts

delete_position_bundle CPI accounts.

DeletePositionBundleCpiBuilder

Instruction builder for DeletePositionBundle via CPI.

DeletePositionBundleInstructionData

DeleteTokenBadge

Accounts.

DeleteTokenBadgeBuilder

Instruction builder for DeleteTokenBadge.

DeleteTokenBadgeCpi

delete_token_badge CPI instruction.

DeleteTokenBadgeCpiAccounts

delete_token_badge CPI accounts.

DeleteTokenBadgeCpiBuilder

Instruction builder for DeleteTokenBadge via CPI.

DeleteTokenBadgeInstructionData

DynamicTickArray

DynamicTickData

FeeTier

FixedTickArray

IdlInclude

Accounts.

IdlIncludeBuilder

Instruction builder for IdlInclude.

IdlIncludeCpi

idl_include CPI instruction.

IdlIncludeCpiAccounts

idl_include CPI accounts.

IdlIncludeCpiBuilder

Instruction builder for IdlInclude via CPI.

IdlIncludeInstructionData

IncreaseLiquidity

Accounts.

IncreaseLiquidityBuilder

Instruction builder for IncreaseLiquidity.

IncreaseLiquidityCpi

increase_liquidity CPI instruction.

IncreaseLiquidityCpiAccounts

increase_liquidity CPI accounts.

IncreaseLiquidityCpiBuilder

Instruction builder for IncreaseLiquidity via CPI.

IncreaseLiquidityInstructionArgs

IncreaseLiquidityInstructionData

IncreaseLiquidityV2

Accounts.

IncreaseLiquidityV2Builder

Instruction builder for IncreaseLiquidityV2.

IncreaseLiquidityV2Cpi

increase_liquidity_v2 CPI instruction.

IncreaseLiquidityV2CpiAccounts

increase_liquidity_v2 CPI accounts.

IncreaseLiquidityV2CpiBuilder

Instruction builder for IncreaseLiquidityV2 via CPI.

IncreaseLiquidityV2InstructionArgs

IncreaseLiquidityV2InstructionData

InitializeAdaptiveFeeTier

Accounts.

InitializeAdaptiveFeeTierBuilder

Instruction builder for InitializeAdaptiveFeeTier.

InitializeAdaptiveFeeTierCpi

initialize_adaptive_fee_tier CPI instruction.

InitializeAdaptiveFeeTierCpiAccounts

initialize_adaptive_fee_tier CPI accounts.

InitializeAdaptiveFeeTierCpiBuilder

Instruction builder for InitializeAdaptiveFeeTier via CPI.

InitializeAdaptiveFeeTierInstructionArgs

InitializeAdaptiveFeeTierInstructionData

InitializeConfig

Accounts.

InitializeConfigBuilder

Instruction builder for InitializeConfig.

InitializeConfigCpi

initialize_config CPI instruction.

InitializeConfigCpiAccounts

initialize_config CPI accounts.

InitializeConfigCpiBuilder

Instruction builder for InitializeConfig via CPI.

InitializeConfigExtension

Accounts.

InitializeConfigExtensionBuilder

Instruction builder for InitializeConfigExtension.

InitializeConfigExtensionCpi

initialize_config_extension CPI instruction.

InitializeConfigExtensionCpiAccounts

initialize_config_extension CPI accounts.

InitializeConfigExtensionCpiBuilder

Instruction builder for InitializeConfigExtension via CPI.

InitializeConfigExtensionInstructionData

InitializeConfigInstructionArgs

InitializeConfigInstructionData

InitializeDynamicTickArray

Accounts.

InitializeDynamicTickArrayBuilder

Instruction builder for InitializeDynamicTickArray.

InitializeDynamicTickArrayCpi

initialize_dynamic_tick_array CPI instruction.

InitializeDynamicTickArrayCpiAccounts

initialize_dynamic_tick_array CPI accounts.

InitializeDynamicTickArrayCpiBuilder

Instruction builder for InitializeDynamicTickArray via CPI.

InitializeDynamicTickArrayInstructionArgs

InitializeDynamicTickArrayInstructionData

InitializeFeeTier

Accounts.

InitializeFeeTierBuilder

Instruction builder for InitializeFeeTier.

InitializeFeeTierCpi

initialize_fee_tier CPI instruction.

InitializeFeeTierCpiAccounts

initialize_fee_tier CPI accounts.

InitializeFeeTierCpiBuilder

Instruction builder for InitializeFeeTier via CPI.

InitializeFeeTierInstructionArgs

InitializeFeeTierInstructionData

InitializePool

Accounts.

InitializePoolBuilder

Instruction builder for InitializePool.

InitializePoolCpi

initialize_pool CPI instruction.

InitializePoolCpiAccounts

initialize_pool CPI accounts.

InitializePoolCpiBuilder

Instruction builder for InitializePool via CPI.

InitializePoolInstructionArgs

InitializePoolInstructionData

InitializePoolV2

Accounts.

InitializePoolV2Builder

Instruction builder for InitializePoolV2.

InitializePoolV2Cpi

initialize_pool_v2 CPI instruction.

InitializePoolV2CpiAccounts

initialize_pool_v2 CPI accounts.

InitializePoolV2CpiBuilder

Instruction builder for InitializePoolV2 via CPI.

InitializePoolV2InstructionArgs

InitializePoolV2InstructionData

InitializePoolWithAdaptiveFee

Accounts.

InitializePoolWithAdaptiveFeeBuilder

Instruction builder for InitializePoolWithAdaptiveFee.

InitializePoolWithAdaptiveFeeCpi

initialize_pool_with_adaptive_fee CPI instruction.

InitializePoolWithAdaptiveFeeCpiAccounts

initialize_pool_with_adaptive_fee CPI accounts.

InitializePoolWithAdaptiveFeeCpiBuilder

Instruction builder for InitializePoolWithAdaptiveFee via CPI.

InitializePoolWithAdaptiveFeeInstructionArgs

InitializePoolWithAdaptiveFeeInstructionData

InitializePositionBundle

Accounts.

InitializePositionBundleBuilder

Instruction builder for InitializePositionBundle.

InitializePositionBundleCpi

initialize_position_bundle CPI instruction.

InitializePositionBundleCpiAccounts

initialize_position_bundle CPI accounts.

InitializePositionBundleCpiBuilder

Instruction builder for InitializePositionBundle via CPI.

InitializePositionBundleInstructionData

InitializePositionBundleWithMetadata

Accounts.

InitializePositionBundleWithMetadataBuilder

Instruction builder for InitializePositionBundleWithMetadata.

InitializePositionBundleWithMetadataCpi

initialize_position_bundle_with_metadata CPI instruction.

InitializePositionBundleWithMetadataCpiAccounts

initialize_position_bundle_with_metadata CPI accounts.

InitializePositionBundleWithMetadataCpiBuilder

Instruction builder for InitializePositionBundleWithMetadata via CPI.

InitializePositionBundleWithMetadataInstructionData

InitializeReward

Accounts.

InitializeRewardBuilder

Instruction builder for InitializeReward.

InitializeRewardCpi

initialize_reward CPI instruction.

InitializeRewardCpiAccounts

initialize_reward CPI accounts.

InitializeRewardCpiBuilder

Instruction builder for InitializeReward via CPI.

InitializeRewardInstructionArgs

InitializeRewardInstructionData

InitializeRewardV2

Accounts.

InitializeRewardV2Builder

Instruction builder for InitializeRewardV2.

InitializeRewardV2Cpi

initialize_reward_v2 CPI instruction.

InitializeRewardV2CpiAccounts

initialize_reward_v2 CPI accounts.

InitializeRewardV2CpiBuilder

Instruction builder for InitializeRewardV2 via CPI.

InitializeRewardV2InstructionArgs

InitializeRewardV2InstructionData

InitializeTickArray

Accounts.

InitializeTickArrayBuilder

Instruction builder for InitializeTickArray.

InitializeTickArrayCpi

initialize_tick_array CPI instruction.

InitializeTickArrayCpiAccounts

initialize_tick_array CPI accounts.

InitializeTickArrayCpiBuilder

Instruction builder for InitializeTickArray via CPI.

InitializeTickArrayInstructionArgs

InitializeTickArrayInstructionData

InitializeTokenBadge

Accounts.

InitializeTokenBadgeBuilder

Instruction builder for InitializeTokenBadge.

InitializeTokenBadgeCpi

initialize_token_badge CPI instruction.

InitializeTokenBadgeCpiAccounts

initialize_token_badge CPI accounts.

InitializeTokenBadgeCpiBuilder

Instruction builder for InitializeTokenBadge via CPI.

InitializeTokenBadgeInstructionData

LiquidityDecreased

LiquidityIncreased

LockConfig

LockPosition

Accounts.

LockPositionBuilder

Instruction builder for LockPosition.

LockPositionCpi

lock_position CPI instruction.

LockPositionCpiAccounts

lock_position CPI accounts.

LockPositionCpiBuilder

Instruction builder for LockPosition via CPI.

LockPositionInstructionArgs

LockPositionInstructionData

MigrateRepurposeRewardAuthoritySpace

Accounts.

MigrateRepurposeRewardAuthoritySpaceBuilder

Instruction builder for MigrateRepurposeRewardAuthoritySpace.

MigrateRepurposeRewardAuthoritySpaceCpi

migrate_repurpose_reward_authority_space CPI instruction.

MigrateRepurposeRewardAuthoritySpaceCpiAccounts

migrate_repurpose_reward_authority_space CPI accounts.

MigrateRepurposeRewardAuthoritySpaceCpiBuilder

Instruction builder for MigrateRepurposeRewardAuthoritySpace via CPI.

MigrateRepurposeRewardAuthoritySpaceInstructionData

OpenBundledPosition

Accounts.

OpenBundledPositionBuilder

Instruction builder for OpenBundledPosition.

OpenBundledPositionCpi

open_bundled_position CPI instruction.

OpenBundledPositionCpiAccounts

open_bundled_position CPI accounts.

OpenBundledPositionCpiBuilder

Instruction builder for OpenBundledPosition via CPI.

OpenBundledPositionInstructionArgs

OpenBundledPositionInstructionData

OpenPosition

Accounts.

OpenPositionBuilder

Instruction builder for OpenPosition.

OpenPositionCpi

open_position CPI instruction.

OpenPositionCpiAccounts

open_position CPI accounts.

OpenPositionCpiBuilder

Instruction builder for OpenPosition via CPI.

OpenPositionInstructionArgs

OpenPositionInstructionData

OpenPositionWithMetadata

Accounts.

OpenPositionWithMetadataBuilder

Instruction builder for OpenPositionWithMetadata.

OpenPositionWithMetadataCpi

open_position_with_metadata CPI instruction.

OpenPositionWithMetadataCpiAccounts

open_position_with_metadata CPI accounts.

OpenPositionWithMetadataCpiBuilder

Instruction builder for OpenPositionWithMetadata via CPI.

OpenPositionWithMetadataInstructionArgs

OpenPositionWithMetadataInstructionData

OpenPositionWithTokenExtensions

Accounts.

OpenPositionWithTokenExtensionsBuilder

Instruction builder for OpenPositionWithTokenExtensions.

OpenPositionWithTokenExtensionsCpi

open_position_with_token_extensions CPI instruction.

OpenPositionWithTokenExtensionsCpiAccounts

open_position_with_token_extensions CPI accounts.

OpenPositionWithTokenExtensionsCpiBuilder

Instruction builder for OpenPositionWithTokenExtensions via CPI.

OpenPositionWithTokenExtensionsInstructionArgs

OpenPositionWithTokenExtensionsInstructionData

Oracle

PoolInitialized

Position

PositionBundle

PositionRewardInfo

RemainingAccountsInfo

RemainingAccountsSlice

ResetPositionRange

Accounts.

ResetPositionRangeBuilder

Instruction builder for ResetPositionRange.

ResetPositionRangeCpi

reset_position_range CPI instruction.

ResetPositionRangeCpiAccounts

reset_position_range CPI accounts.

ResetPositionRangeCpiBuilder

Instruction builder for ResetPositionRange via CPI.

ResetPositionRangeInstructionArgs

ResetPositionRangeInstructionData

SetCollectProtocolFeesAuthority

Accounts.

SetCollectProtocolFeesAuthorityBuilder

Instruction builder for SetCollectProtocolFeesAuthority.

SetCollectProtocolFeesAuthorityCpi

set_collect_protocol_fees_authority CPI instruction.

SetCollectProtocolFeesAuthorityCpiAccounts

set_collect_protocol_fees_authority CPI accounts.

SetCollectProtocolFeesAuthorityCpiBuilder

Instruction builder for SetCollectProtocolFeesAuthority via CPI.

SetCollectProtocolFeesAuthorityInstructionData

SetConfigExtensionAuthority

Accounts.

SetConfigExtensionAuthorityBuilder

Instruction builder for SetConfigExtensionAuthority.

SetConfigExtensionAuthorityCpi

set_config_extension_authority CPI instruction.

SetConfigExtensionAuthorityCpiAccounts

set_config_extension_authority CPI accounts.

SetConfigExtensionAuthorityCpiBuilder

Instruction builder for SetConfigExtensionAuthority via CPI.

SetConfigExtensionAuthorityInstructionData

SetConfigFeatureFlag

Accounts.

SetConfigFeatureFlagBuilder

Instruction builder for SetConfigFeatureFlag.

SetConfigFeatureFlagCpi

set_config_feature_flag CPI instruction.

SetConfigFeatureFlagCpiAccounts

set_config_feature_flag CPI accounts.

SetConfigFeatureFlagCpiBuilder

Instruction builder for SetConfigFeatureFlag via CPI.

SetConfigFeatureFlagInstructionArgs

SetConfigFeatureFlagInstructionData

SetDefaultBaseFeeRate

Accounts.

SetDefaultBaseFeeRateBuilder

Instruction builder for SetDefaultBaseFeeRate.

SetDefaultBaseFeeRateCpi

set_default_base_fee_rate CPI instruction.

SetDefaultBaseFeeRateCpiAccounts

set_default_base_fee_rate CPI accounts.

SetDefaultBaseFeeRateCpiBuilder

Instruction builder for SetDefaultBaseFeeRate via CPI.

SetDefaultBaseFeeRateInstructionArgs

SetDefaultBaseFeeRateInstructionData

SetDefaultFeeRate

Accounts.

SetDefaultFeeRateBuilder

Instruction builder for SetDefaultFeeRate.

SetDefaultFeeRateCpi

set_default_fee_rate CPI instruction.

SetDefaultFeeRateCpiAccounts

set_default_fee_rate CPI accounts.

SetDefaultFeeRateCpiBuilder

Instruction builder for SetDefaultFeeRate via CPI.

SetDefaultFeeRateInstructionArgs

SetDefaultFeeRateInstructionData

SetDefaultProtocolFeeRate

Accounts.

SetDefaultProtocolFeeRateBuilder

Instruction builder for SetDefaultProtocolFeeRate.

SetDefaultProtocolFeeRateCpi

set_default_protocol_fee_rate CPI instruction.

SetDefaultProtocolFeeRateCpiAccounts

set_default_protocol_fee_rate CPI accounts.

SetDefaultProtocolFeeRateCpiBuilder

Instruction builder for SetDefaultProtocolFeeRate via CPI.

SetDefaultProtocolFeeRateInstructionArgs

SetDefaultProtocolFeeRateInstructionData

SetDelegatedFeeAuthority

Accounts.

SetDelegatedFeeAuthorityBuilder

Instruction builder for SetDelegatedFeeAuthority.

SetDelegatedFeeAuthorityCpi

set_delegated_fee_authority CPI instruction.

SetDelegatedFeeAuthorityCpiAccounts

set_delegated_fee_authority CPI accounts.

SetDelegatedFeeAuthorityCpiBuilder

Instruction builder for SetDelegatedFeeAuthority via CPI.

SetDelegatedFeeAuthorityInstructionData

SetFeeAuthority

Accounts.

SetFeeAuthorityBuilder

Instruction builder for SetFeeAuthority.

SetFeeAuthorityCpi

set_fee_authority CPI instruction.

SetFeeAuthorityCpiAccounts

set_fee_authority CPI accounts.

SetFeeAuthorityCpiBuilder

Instruction builder for SetFeeAuthority via CPI.

SetFeeAuthorityInstructionData

SetFeeRate

Accounts.

SetFeeRateBuilder

Instruction builder for SetFeeRate.

SetFeeRateByDelegatedFeeAuthority

Accounts.

SetFeeRateByDelegatedFeeAuthorityBuilder

Instruction builder for SetFeeRateByDelegatedFeeAuthority.

SetFeeRateByDelegatedFeeAuthorityCpi

set_fee_rate_by_delegated_fee_authority CPI instruction.

SetFeeRateByDelegatedFeeAuthorityCpiAccounts

set_fee_rate_by_delegated_fee_authority CPI accounts.

SetFeeRateByDelegatedFeeAuthorityCpiBuilder

Instruction builder for SetFeeRateByDelegatedFeeAuthority via CPI.

SetFeeRateByDelegatedFeeAuthorityInstructionArgs

SetFeeRateByDelegatedFeeAuthorityInstructionData

SetFeeRateCpi

set_fee_rate CPI instruction.

SetFeeRateCpiAccounts

set_fee_rate CPI accounts.

SetFeeRateCpiBuilder

Instruction builder for SetFeeRate via CPI.

SetFeeRateInstructionArgs

SetFeeRateInstructionData

SetInitializePoolAuthority

Accounts.

SetInitializePoolAuthorityBuilder

Instruction builder for SetInitializePoolAuthority.

SetInitializePoolAuthorityCpi

set_initialize_pool_authority CPI instruction.

SetInitializePoolAuthorityCpiAccounts

set_initialize_pool_authority CPI accounts.

SetInitializePoolAuthorityCpiBuilder

Instruction builder for SetInitializePoolAuthority via CPI.

SetInitializePoolAuthorityInstructionData

SetPresetAdaptiveFeeConstants

Accounts.

SetPresetAdaptiveFeeConstantsBuilder

Instruction builder for SetPresetAdaptiveFeeConstants.

SetPresetAdaptiveFeeConstantsCpi

set_preset_adaptive_fee_constants CPI instruction.

SetPresetAdaptiveFeeConstantsCpiAccounts

set_preset_adaptive_fee_constants CPI accounts.

SetPresetAdaptiveFeeConstantsCpiBuilder

Instruction builder for SetPresetAdaptiveFeeConstants via CPI.

SetPresetAdaptiveFeeConstantsInstructionArgs

SetPresetAdaptiveFeeConstantsInstructionData

SetProtocolFeeRate

Accounts.

SetProtocolFeeRateBuilder

Instruction builder for SetProtocolFeeRate.

SetProtocolFeeRateCpi

set_protocol_fee_rate CPI instruction.

SetProtocolFeeRateCpiAccounts

set_protocol_fee_rate CPI accounts.

SetProtocolFeeRateCpiBuilder

Instruction builder for SetProtocolFeeRate via CPI.

SetProtocolFeeRateInstructionArgs

SetProtocolFeeRateInstructionData

SetRewardAuthority

Accounts.

SetRewardAuthorityBuilder

Instruction builder for SetRewardAuthority.

SetRewardAuthorityBySuperAuthority

Accounts.

SetRewardAuthorityBySuperAuthorityBuilder

Instruction builder for SetRewardAuthorityBySuperAuthority.

SetRewardAuthorityBySuperAuthorityCpi

set_reward_authority_by_super_authority CPI instruction.

SetRewardAuthorityBySuperAuthorityCpiAccounts

set_reward_authority_by_super_authority CPI accounts.

SetRewardAuthorityBySuperAuthorityCpiBuilder

Instruction builder for SetRewardAuthorityBySuperAuthority via CPI.

SetRewardAuthorityBySuperAuthorityInstructionArgs

SetRewardAuthorityBySuperAuthorityInstructionData

SetRewardAuthorityCpi

set_reward_authority CPI instruction.

SetRewardAuthorityCpiAccounts

set_reward_authority CPI accounts.

SetRewardAuthorityCpiBuilder

Instruction builder for SetRewardAuthority via CPI.

SetRewardAuthorityInstructionArgs

SetRewardAuthorityInstructionData

SetRewardEmissions

Accounts.

SetRewardEmissionsBuilder

Instruction builder for SetRewardEmissions.

SetRewardEmissionsCpi

set_reward_emissions CPI instruction.

SetRewardEmissionsCpiAccounts

set_reward_emissions CPI accounts.

SetRewardEmissionsCpiBuilder

Instruction builder for SetRewardEmissions via CPI.

SetRewardEmissionsInstructionArgs

SetRewardEmissionsInstructionData

SetRewardEmissionsSuperAuthority

Accounts.

SetRewardEmissionsSuperAuthorityBuilder

Instruction builder for SetRewardEmissionsSuperAuthority.

SetRewardEmissionsSuperAuthorityCpi

set_reward_emissions_super_authority CPI instruction.

SetRewardEmissionsSuperAuthorityCpiAccounts

set_reward_emissions_super_authority CPI accounts.

SetRewardEmissionsSuperAuthorityCpiBuilder

Instruction builder for SetRewardEmissionsSuperAuthority via CPI.

SetRewardEmissionsSuperAuthorityInstructionData

SetRewardEmissionsV2

Accounts.

SetRewardEmissionsV2Builder

Instruction builder for SetRewardEmissionsV2.

SetRewardEmissionsV2Cpi

set_reward_emissions_v2 CPI instruction.

SetRewardEmissionsV2CpiAccounts

set_reward_emissions_v2 CPI accounts.

SetRewardEmissionsV2CpiBuilder

Instruction builder for SetRewardEmissionsV2 via CPI.

SetRewardEmissionsV2InstructionArgs

SetRewardEmissionsV2InstructionData

SetTokenBadgeAttribute

Accounts.

SetTokenBadgeAttributeBuilder

Instruction builder for SetTokenBadgeAttribute.

SetTokenBadgeAttributeCpi

set_token_badge_attribute CPI instruction.

SetTokenBadgeAttributeCpiAccounts

set_token_badge_attribute CPI accounts.

SetTokenBadgeAttributeCpiBuilder

Instruction builder for SetTokenBadgeAttribute via CPI.

SetTokenBadgeAttributeInstructionArgs

SetTokenBadgeAttributeInstructionData

SetTokenBadgeAuthority

Accounts.

SetTokenBadgeAuthorityBuilder

Instruction builder for SetTokenBadgeAuthority.

SetTokenBadgeAuthorityCpi

set_token_badge_authority CPI instruction.

SetTokenBadgeAuthorityCpiAccounts

set_token_badge_authority CPI accounts.

SetTokenBadgeAuthorityCpiBuilder

Instruction builder for SetTokenBadgeAuthority via CPI.

SetTokenBadgeAuthorityInstructionData

Swap

Accounts.

SwapBuilder

Instruction builder for Swap.

SwapCpi

swap CPI instruction.

SwapCpiAccounts

swap CPI accounts.

SwapCpiBuilder

Instruction builder for Swap via CPI.

SwapInstructionArgs

SwapInstructionData

SwapV2

Accounts.

SwapV2Builder

Instruction builder for SwapV2.

SwapV2Cpi

swap_v2 CPI instruction.

SwapV2CpiAccounts

swap_v2 CPI accounts.

SwapV2CpiBuilder

Instruction builder for SwapV2 via CPI.

SwapV2InstructionArgs

SwapV2InstructionData

Tick

TokenBadge

Traded

TransferLockedPosition

Accounts.

TransferLockedPositionBuilder

Instruction builder for TransferLockedPosition.

TransferLockedPositionCpi

transfer_locked_position CPI instruction.

TransferLockedPositionCpiAccounts

transfer_locked_position CPI accounts.

TransferLockedPositionCpiBuilder

Instruction builder for TransferLockedPosition via CPI.

TransferLockedPositionInstructionData

TwoHopSwap

Accounts.

TwoHopSwapBuilder

Instruction builder for TwoHopSwap.

TwoHopSwapCpi

two_hop_swap CPI instruction.

TwoHopSwapCpiAccounts

two_hop_swap CPI accounts.

TwoHopSwapCpiBuilder

Instruction builder for TwoHopSwap via CPI.

TwoHopSwapInstructionArgs

TwoHopSwapInstructionData

TwoHopSwapV2

Accounts.

TwoHopSwapV2Builder

Instruction builder for TwoHopSwapV2.

TwoHopSwapV2Cpi

two_hop_swap_v2 CPI instruction.

TwoHopSwapV2CpiAccounts

two_hop_swap_v2 CPI accounts.

TwoHopSwapV2CpiBuilder

Instruction builder for TwoHopSwapV2 via CPI.

TwoHopSwapV2InstructionArgs

TwoHopSwapV2InstructionData

UpdateFeesAndRewards

Accounts.

UpdateFeesAndRewardsBuilder

Instruction builder for UpdateFeesAndRewards.

UpdateFeesAndRewardsCpi

update_fees_and_rewards CPI instruction.

UpdateFeesAndRewardsCpiAccounts

update_fees_and_rewards CPI accounts.

UpdateFeesAndRewardsCpiBuilder

Instruction builder for UpdateFeesAndRewards via CPI.

UpdateFeesAndRewardsInstructionData

Whirlpool

WhirlpoolRewardInfo

Stores the state relevant for tracking liquidity mining rewards at the Whirlpool level. These values are used in conjunction with PositionRewardInfo, Tick.reward_growths_outside, and Whirlpool.reward_last_updated_timestamp to determine how many rewards are earned by open positions.

WhirlpoolsConfig

WhirlpoolsConfigExtension

Enums

AccountsType

AdaptiveFeeTierFilter

ConfigFeatureFlag

DynamicTick

DynamicTickArrayFilter

FeeTierFilter

FixedTickArrayFilter

LockConfigFilter

LockType

LockTypeLabel

MaybeAccount

OracleFilter

PositionBundleFilter

PositionFilter

TickArray

TickArrayFilter

TokenBadgeAttribute

TokenBadgeFilter

WhirlpoolError

WhirlpoolFilter

WhirlpoolsConfigExtensionFilter

WhirlpoolsConfigFilter

Constants

ADAPTIVE_FEE_TIER_DISCRIMINATOR

CLOSE_BUNDLED_POSITION_DISCRIMINATOR

CLOSE_POSITION_DISCRIMINATOR

CLOSE_POSITION_WITH_TOKEN_EXTENSIONS_DISCRIMINATOR

COLLECT_FEES_DISCRIMINATOR

COLLECT_FEES_V2_DISCRIMINATOR

COLLECT_PROTOCOL_FEES_DISCRIMINATOR

COLLECT_PROTOCOL_FEES_V2_DISCRIMINATOR

COLLECT_REWARD_DISCRIMINATOR

COLLECT_REWARD_V2_DISCRIMINATOR

DECREASE_LIQUIDITY_DISCRIMINATOR

DECREASE_LIQUIDITY_V2_DISCRIMINATOR

DELETE_POSITION_BUNDLE_DISCRIMINATOR

DELETE_TOKEN_BADGE_DISCRIMINATOR

DYNAMIC_TICK_ARRAY_DISCRIMINATOR

FEE_TIER_DISCRIMINATOR

FIXED_TICK_ARRAY_DISCRIMINATOR

ID

whirlpool program ID.

IDL_INCLUDE_DISCRIMINATOR

INCREASE_LIQUIDITY_DISCRIMINATOR

INCREASE_LIQUIDITY_V2_DISCRIMINATOR

INITIALIZE_ADAPTIVE_FEE_TIER_DISCRIMINATOR

INITIALIZE_CONFIG_DISCRIMINATOR

INITIALIZE_CONFIG_EXTENSION_DISCRIMINATOR

INITIALIZE_DYNAMIC_TICK_ARRAY_DISCRIMINATOR

INITIALIZE_FEE_TIER_DISCRIMINATOR

INITIALIZE_POOL_DISCRIMINATOR

INITIALIZE_POOL_V2_DISCRIMINATOR

INITIALIZE_POOL_WITH_ADAPTIVE_FEE_DISCRIMINATOR

INITIALIZE_POSITION_BUNDLE_DISCRIMINATOR

INITIALIZE_POSITION_BUNDLE_WITH_METADATA_DISCRIMINATOR

INITIALIZE_REWARD_DISCRIMINATOR

INITIALIZE_REWARD_V2_DISCRIMINATOR

INITIALIZE_TICK_ARRAY_DISCRIMINATOR

INITIALIZE_TOKEN_BADGE_DISCRIMINATOR

LOCK_CONFIG_DISCRIMINATOR

LOCK_POSITION_DISCRIMINATOR

MIGRATE_REPURPOSE_REWARD_AUTHORITY_SPACE_DISCRIMINATOR

OPEN_BUNDLED_POSITION_DISCRIMINATOR

OPEN_POSITION_DISCRIMINATOR

OPEN_POSITION_WITH_METADATA_DISCRIMINATOR

OPEN_POSITION_WITH_TOKEN_EXTENSIONS_DISCRIMINATOR

ORACLE_DISCRIMINATOR

POSITION_BUNDLE_DISCRIMINATOR

POSITION_DISCRIMINATOR

RESET_POSITION_RANGE_DISCRIMINATOR

SET_COLLECT_PROTOCOL_FEES_AUTHORITY_DISCRIMINATOR

SET_CONFIG_EXTENSION_AUTHORITY_DISCRIMINATOR

SET_CONFIG_FEATURE_FLAG_DISCRIMINATOR

SET_DEFAULT_BASE_FEE_RATE_DISCRIMINATOR

SET_DEFAULT_FEE_RATE_DISCRIMINATOR

SET_DEFAULT_PROTOCOL_FEE_RATE_DISCRIMINATOR

SET_DELEGATED_FEE_AUTHORITY_DISCRIMINATOR

SET_FEE_AUTHORITY_DISCRIMINATOR

SET_FEE_RATE_BY_DELEGATED_FEE_AUTHORITY_DISCRIMINATOR

SET_FEE_RATE_DISCRIMINATOR

SET_INITIALIZE_POOL_AUTHORITY_DISCRIMINATOR

SET_PRESET_ADAPTIVE_FEE_CONSTANTS_DISCRIMINATOR

SET_PROTOCOL_FEE_RATE_DISCRIMINATOR

SET_REWARD_AUTHORITY_BY_SUPER_AUTHORITY_DISCRIMINATOR

SET_REWARD_AUTHORITY_DISCRIMINATOR

SET_REWARD_EMISSIONS_DISCRIMINATOR

SET_REWARD_EMISSIONS_SUPER_AUTHORITY_DISCRIMINATOR

SET_REWARD_EMISSIONS_V2_DISCRIMINATOR

SET_TOKEN_BADGE_ATTRIBUTE_DISCRIMINATOR

SET_TOKEN_BADGE_AUTHORITY_DISCRIMINATOR

SWAP_DISCRIMINATOR

SWAP_V2_DISCRIMINATOR

TOKEN_BADGE_DISCRIMINATOR

TRANSFER_LOCKED_POSITION_DISCRIMINATOR

TWO_HOP_SWAP_DISCRIMINATOR

TWO_HOP_SWAP_V2_DISCRIMINATOR

UPDATE_FEES_AND_REWARDS_DISCRIMINATOR

WHIRLPOOLS_CONFIG_DISCRIMINATOR

WHIRLPOOLS_CONFIG_EXTENSION_DISCRIMINATOR

WHIRLPOOL_DISCRIMINATOR

WHIRLPOOL_ID

whirlpool program ID.

Functions

fetch_adaptive_fee_tier

fetch_all_adaptive_fee_tier

fetch_all_adaptive_fee_tier_with_filter

fetch_all_dynamic_tick_array

fetch_all_dynamic_tick_array_with_filter

fetch_all_fee_tier

fetch_all_fee_tier_with_filter

fetch_all_fixed_tick_array

fetch_all_fixed_tick_array_with_filter

fetch_all_lock_config

fetch_all_lock_config_with_filter

fetch_all_maybe_adaptive_fee_tier

fetch_all_maybe_dynamic_tick_array

fetch_all_maybe_fee_tier

fetch_all_maybe_fixed_tick_array

fetch_all_maybe_lock_config

fetch_all_maybe_oracle

fetch_all_maybe_position

fetch_all_maybe_position_bundle

fetch_all_maybe_tick_array

fetch_all_maybe_token_badge

fetch_all_maybe_whirlpool

fetch_all_maybe_whirlpools_config

fetch_all_maybe_whirlpools_config_extension

fetch_all_oracle

fetch_all_oracle_with_filter

fetch_all_position

fetch_all_position_bundle

fetch_all_position_bundle_with_filter

fetch_all_position_with_filter

fetch_all_tick_array

fetch_all_tick_array_with_filter

fetch_all_token_badge

fetch_all_token_badge_with_filter

fetch_all_whirlpool

fetch_all_whirlpool_with_filter

fetch_all_whirlpools_config

fetch_all_whirlpools_config_extension

fetch_all_whirlpools_config_extension_with_filter

fetch_all_whirlpools_config_with_filter

fetch_dynamic_tick_array

fetch_fee_tier

fetch_fixed_tick_array

fetch_lock_config

fetch_maybe_adaptive_fee_tier

fetch_maybe_dynamic_tick_array

fetch_maybe_fee_tier

fetch_maybe_fixed_tick_array

fetch_maybe_lock_config

fetch_maybe_oracle

fetch_maybe_position

fetch_maybe_position_bundle

fetch_maybe_tick_array

fetch_maybe_token_badge

fetch_maybe_whirlpool

fetch_maybe_whirlpools_config

fetch_maybe_whirlpools_config_extension

fetch_oracle

fetch_position

fetch_position_bundle

fetch_tick_array

fetch_token_badge

fetch_whirlpool

fetch_whirlpools_config

fetch_whirlpools_config_extension

get_bundled_position_address

get_fee_tier_address

get_lock_config_address

get_oracle_address

get_position_address

get_position_bundle_address

get_tick_array_address

get_token_badge_address

get_whirlpool_address

get_whirlpools_config_extension_address