orca_whirlpools_docs

Module orca_whirlpools_core

Orca Whirlpools Core SDK

This package provides developers with advanced functionalities for interacting with the Whirlpool Program on Solana. The Core SDK offers convenient methods for math calculations, quotes, and other utilities, making it easier to manage complex liquidity and swap operations within Rust projects. It serves as the foundation for the High-Level SDK, providing key building blocks that enable developers to perform sophisticated actions while maintaining control over core details.

Key Features

Math Library: Contains a variety of functions for math operations related to bundles, positions, prices, ticks, and tokens, including calculations such as determining position status or price conversions.

Quote Library: Provides utility functions for generating quotes, such as increasing liquidity, collecting fees or rewards, and swapping, enabling precise and optimized decision-making in liquidity management.

Installation

cargo add orca_whirlpools_core

Usage

Here are some basic examples of how to use the package:

Math Example

The following example demonstrates how to use the is_position_in_range function to determine whether a position is currently in range.

use orca_whirlpools_core::is_position_in_range;

fn main() {
    let current_sqrt_price = 7448043534253661173u128;
    let tick_index_1 = -18304;
    let tick_index_2 = -17956;

    let in_range = is_position_in_range(current_sqrt_price.into(), tick_index_1, tick_index_2);
    println!("Position in range? {:?}", in_range);
}

Expected output:

Position in range? true

Adjust Liquidity Quote Example

The following example demonstrates how to use the increase_liquidity_quote_a function to calculate a quote for increasing liquidity given a token A amount.

use orca_whirlpools_core::increase_liquidity_quote_a;
use orca_whirlpools_core::TransferFee;

fn main() {
    let token_amount_a = 1000000000u64;
    let slippage_tolerance_bps = 100u16;
    let current_sqrt_price = 7437568627975669726u128;
    let tick_index_1 = -18568;
    let tick_index_2 = -17668;
    let transfer_fee_a = Some(TransferFee::new(200));
    let transfer_fee_b = None;

    let quote = increase_liquidity_quote_a(
        token_amount_a,
        slippage_tolerance_bps,
        current_sqrt_price.into(),
        tick_index_1,
        tick_index_2,
        transfer_fee_a,
        transfer_fee_b,
    ).unwrap();
    
    println!("{:?}", quote);
}

Expected output:

IncreaseLiquidityQuote {
    liquidity_delta: 16011047470,
    token_est_a: 1000000000,
    token_est_b: 127889169,
    token_max_a: 1010000000,
    token_max_b: 129168061,
}

Structs

AdaptiveFeeConstantsFacade

AdaptiveFeeInfo

AdaptiveFeeVariablesFacade

CollectFeesQuote

CollectRewardQuote

CollectRewardsQuote

DecreaseLiquidityQuote

ExactInSwapQuote

ExactOutSwapQuote

IncreaseLiquidityQuote

OracleFacade

PositionFacade

PositionRatio

PositionRewardInfoFacade

SwapResult

TickArrayFacade

TickArraySequence

TickFacade

TickRange

TransferFee

WhirlpoolFacade

WhirlpoolRewardInfoFacade

Enums

PositionStatus

TickArrays

Constants

ADAPTIVE_FEE_CONTROL_FACTOR_DENOMINATOR

adaptive_fee_control_factor is used to map the square of the volatility accumulator to the fee rate.

AMOUNT_EXCEEDS_MAX_U64

ARITHMETIC_OVERFLOW

BPS_DENOMINATOR

FEE_RATE_DENOMINATOR

The denominator of the fee rate value.

FEE_RATE_HARD_LIMIT

max fee rate should be controlled by max_volatility_accumulator, so this is a hard limit for safety. Fee rate is represented as hundredths of a basis point.

FULL_RANGE_ONLY_TICK_SPACING_THRESHOLD

Pools with tick spacing above this threshold are considered full range only. This means the program rejects any non-full range positions in these pools.

INVALID_ADAPTIVE_FEE_INFO

INVALID_SLIPPAGE_TOLERANCE

INVALID_SQRT_PRICE_LIMIT_DIRECTION

INVALID_TICK_ARRAY_SEQUENCE

INVALID_TICK_INDEX

INVALID_TIMESTAMP

INVALID_TRANSFER_FEE

MAX_REFERENCE_AGE

The time (in seconds) to forcibly reset the reference if it is not updated for a long time.

MAX_SQRT_PRICE

The maximum sqrt price for a whirlpool.

MAX_TICK_INDEX

The maximum tick index.

MIN_SQRT_PRICE

The minimum sqrt price for a whirlpool.

MIN_TICK_INDEX

The minimum tick index.

NUM_REWARDS

The number of reward tokens in a pool.

POSITION_BUNDLE_SIZE

The maximum number of positions in a position bundle.

REDUCTION_FACTOR_DENOMINATOR

The denominator of the reduction factor.

SQRT_PRICE_LIMIT_OUT_OF_BOUNDS

SQRT_PRICE_OUT_OF_BOUNDS

TICK_ARRAY_NOT_EVENLY_SPACED

TICK_ARRAY_SIZE

The number of ticks in a tick array.

TICK_INDEX_NOT_IN_ARRAY

TICK_INDEX_OUT_OF_BOUNDS

TICK_SEQUENCE_EMPTY

VOLATILITY_ACCUMULATOR_SCALE_FACTOR

This constant is used to scale the value of the volatility accumulator.

ZERO_TRADABLE_AMOUNT

Functions

collect_fees_quote

Calculate fees owed for a position

collect_rewards_quote

Calculate rewards owed for a position

compute_swap

Computes the amounts of tokens A and B based on the current Whirlpool state and tick sequence.

decrease_liquidity_quote

Calculate the quote for decreasing liquidity

decrease_liquidity_quote_a

Calculate the quote for decreasing liquidity given a token a amount

decrease_liquidity_quote_b

Calculate the quote for decreasing liquidity given a token b amount

first_unoccupied_position_in_bundle

Get the first unoccupied position in a bundle

get_full_range_tick_indexes

Get the minimum and maximum tick index that can be initialized.

get_initializable_tick_index

Get the initializable tick index. If the tick index is already initializable, it is returned as is.

get_next_initializable_tick_index

Get the next initializable tick index.

get_prev_initializable_tick_index

Get the previous initializable tick index.

get_tick_array_start_tick_index

Get the first tick index in the tick array that contains the specified tick index.

get_tick_index_in_array

Get the index of a tick in a tick array.

increase_liquidity_quote

Calculate the quote for increasing liquidity

increase_liquidity_quote_a

Calculate the quote for increasing liquidity given a token a amount

increase_liquidity_quote_b

Calculate the quote for increasing liquidity given a token b amount

invert_price

Invert a price IMPORTANT: floating point operations can reduce the precision of the result. Make sure to do these operations last and not to use the result for further calculations.

invert_sqrt_price

Get the sqrt price for the inverse of the price that this tick represents. Because converting to a tick index and then back to a sqrt price is lossy, this function is clamped to the nearest tick index.

invert_tick_index

Get the tick index for the inverse of the price that this tick represents. Eg: Consider tick i where Pb/Pa = 1.0001 ^ i inverse of this, i.e. Pa/Pb = 1 / (1.0001 ^ i) = 1.0001^-i

is_full_range_only

Check if a whirlpool is a full-range only pool.

is_initialized_with_adaptive_fee

Check if the whirlpool is initialized with adaptive fee

is_position_bundle_empty

Check whether a position bundle is empty

is_position_bundle_full

Check whether a position bundle is full A position bundle can contain 256 positions

is_position_in_range

Check if a position is in range. When a position is in range it is earning fees and rewards

is_tick_index_in_bounds

Check if a tick is in-bounds.

is_tick_initializable

Check if a tick is initializable. A tick is initializable if it is divisible by the tick spacing.

order_tick_indexes

Order tick indexes in ascending order. If the lower tick index is greater than the upper tick index, the indexes are swapped. This is useful for ensuring that the lower tick index is always less than the upper tick index.

position_ratio

Calculate the token_a / token_b ratio of a (ficticious) position

position_status

Calculate the status of a position The status can be one of three values:

price_to_sqrt_price

Convert a price into a sqrt priceX64 IMPORTANT: floating point operations can reduce the precision of the result. Make sure to do these operations last and not to use the result for further calculations.

price_to_tick_index

Convert a price into a tick index IMPORTANT: floating point operations can reduce the precision of the result. Make sure to do these operations last and not to use the result for further calculations.

sqrt_price_to_price

Convert a sqrt priceX64 into a tick index IMPORTANT: floating point operations can reduce the precision of the result. Make sure to do these operations last and not to use the result for further calculations.

sqrt_price_to_tick_index

Derive the tick index from a sqrt price. The precision of this method is only guarranted if tick is within the bounds of {max, min} tick-index. This function will make panic for zero sqrt price.

swap_quote_by_input_token

Computes the exact input or output amount for a swap transaction.

swap_quote_by_output_token

Computes the exact input or output amount for a swap transaction.

tick_index_to_price

Convert a tick index into a price IMPORTANT: floating point operations can reduce the precision of the result. Make sure to do these operations last and not to use the result for further calculations.

tick_index_to_sqrt_price

Derive the sqrt-price from a tick index. The precision of this method is only guarranted if tick is within the bounds of {max, min} tick-index.

try_apply_swap_fee

Apply a swap fee to an amount e.g. You send 10000 amount with 10000 fee rate. The fee amount will be 100. So the amount after fee will be 9900.

try_apply_transfer_fee

Apply a transfer fee to an amount e.g. You send 10000 amount with 100 fee rate. The fee amount will be 100. So the amount after fee will be 9900.

try_get_amount_delta_a

Calculate the amount A delta between two sqrt_prices

try_get_amount_delta_b

Calculate the amount B delta between two sqrt_prices

try_get_max_amount_with_slippage_tolerance

Get the maximum amount with a slippage tolerance e.g. Your estimated amount you send is 10000 with 100 slippage tolerance. The max you send will be 10100.

try_get_min_amount_with_slippage_tolerance

Get the minimum amount with a slippage tolerance e.g. Your estimated amount you receive is 10000 with 100 slippage tolerance. The min amount you receive will be 9900.

try_get_next_sqrt_price_from_a

Calculate the next square root price

try_get_next_sqrt_price_from_b

Calculate the next square root price

try_get_token_estimates_from_liquidity

Calculate the estimated token amounts for a given liquidity delta and price range

try_reverse_apply_swap_fee

Reverse the application of a swap fee to an amount e.g. You received 9900 amount with 10000 fee rate. The fee amount will be 100. So the amount before fee will be 10000.

try_reverse_apply_transfer_fee

Reverse the application of a transfer fee to an amount e.g. You received 9900 amount with 100 fee rate. The fee amount will be 100. So the amount before fee will be 10000.

Type Aliases

CoreError

U128